+++ R/CytoProcessingStepImplementations.R
@@ -55,8 +55,8 @@
 estimateScaleTransforms <- function(ff,
                                     fluoMethod = c("estimateLogicle",
                                                    "none"),
-                                    scatterMethod = c("linearQuantile", 
-                                                      "none"),
+                                    scatterMethod = c("none",
+                                                      "linearQuantile"),
                                     scatterRefMarker = NULL,
                                     specificScatterChannels = NULL){
     fluoMethod <- match.arg(fluoMethod)